Fix flickering of list items on scrollIntoView
We cannot use the react-bootstrap `ListGroup.Item` here, because it sets the `active` class automatically at `onClick`. This behavior cannot be supressed and interfers with the Bootstrap scrollspy + `scrollIntoView` which causes a flickering of the navigation items. The only solution is to use a plain `a` element and add the necessary Bootstrap classes here.